add support for overflow exception

+++ Calculator/Result.fs
@@ -1,7 +1,7 @@
 ﻿namespace Calculator
 
 type Result =
-    | Integer of Value : int64
+    | Integer of Value : int32
     | Real of Value : double
     | String of Text : string
     | Error
